requirements.txt
gevent
gevent-websocket
greenlet
proxytypes3
nose
feedparser
Jinja2
Werkzeug
Flask
Flask-Cors
redis
psycopg2-binary
msgpack
python-dateutil
gunicorn
bugsnag
blinker
regex
Flask-Login
Flask-WTF
WTForms
email-validator
werkzeug
python-dotenv
flask-socketio
simple-websocket
Flask-SocketIO
